api/pyproject.toml
[project]
name = "todo-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "faker>=37.3.0",
    "fastapi-camelcase>=2.0.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-watcher>=0.4.3",
    "sqlalchemy>=2.0.41",
]
